fix build error

+++ tools/analyzer/source.cpp
@@ -101,7 +101,7 @@ Optional<Range> Source::toRange(Token token) const {
 
 SourcePtr SourceManager::findById(ModId id) const {
   auto v = toUnderlying(id);
-  if (v > 0 && v - 1 < this->entries.size()) {
+  if (v > 0 && static_cast<unsigned int>(v - 1) < this->entries.size()) {
     return this->entries[v - 1];
   }
   return nullptr;
